Move comment type definitions to CommentTypes.ts.
