--- src/model/types/CommentType.ts
@@ -1,34 +1,0 @@
-export enum CommentType {
-  ACTIVITY_REGULATION = 'ACTIVITY REGULATION',
-  ALLERGEN = 'ALLERGEN',
-  ALTERNATIVE_PRODUCTS = 'ALTERNATIVE PRODUCTS',
-  BIOPHYSICOCHEMICAL_PROPERTIES = 'BIOPHYSICOCHEMICAL PROPERTIES',
-  BIOTECHNOLOGY = 'BIOTECHNOLOGY',
-  CATALYTIC_ACTIVITY = 'CATALYTIC ACTIVITY',
-  CAUTION = 'CAUTION',
-  COFACTOR = 'COFACTOR',
-  DEVELOPMENTAL_STAGE = 'DEVELOPMENTAL STAGE',
-  DISEASE = 'DISEASE',
-  DISRUPTION_PHENOTYPE = 'DISRUPTION PHENOTYPE',
-  DOMAIN = 'DOMAIN',
-  FUNCTION = 'FUNCTION',
-  INDUCTION = 'INDUCTION',
-  INTERACTION = 'INTERACTION',
-  MASS_SPECTROMETRY = 'MASS SPECTROMETRY',
-  MISCELLANEOUS = 'MISCELLANEOUS',
-  PATHWAY = 'PATHWAY',
-  PHARMACEUTICAL = 'PHARMACEUTICAL',
-  POLYMORPHISM = 'POLYMORPHISM',
-  PTM = 'PTM',
-  RNA_EDITING = 'RNA EDITING',
-  SEQUENCE_CAUTION = 'SEQUENCE CAUTION',
-  SIMILARITY = 'SIMILARITY',
-  SUBCELLULAR_LOCATION = 'SUBCELLULAR LOCATION',
-  SUBUNIT = 'SUBUNIT',
-  TISSUE_SPECIFICITY = 'TISSUE SPECIFICITY',
-  TOXIC_DOSE = 'TOXIC DOSE',
-  UNKNOWN = 'UNKNOWN',
-  WEBRESOURCE = 'WEB RESOURCE',
-}
-
-export default CommentType;
