+++ src/model/uniprotkb/sections/ExpressionConverter.ts
@@ -1,4 +1,4 @@
-import CommentType from '../../types/CommentType';
+import { CommentType } from '../../types/CommentTypes';
 import KeywordCategory from '../../types/KeywordCategory';
 import EntrySection from '../../types/EntrySection';
 import { convertSection } from '../SectionConverter';
@@ -6,7 +6,10 @@ import { UniProtkbAPIModel } from '../UniProtkbConverter';
 
 const keywordsCategories = [KeywordCategory.DEVELOPMENTAL_STAGE];
 
-const commentsCategories = [CommentType.TISSUE_SPECIFICITY, CommentType.INDUCTION];
+const commentsCategories = [
+  CommentType.TISSUE_SPECIFICITY,
+  CommentType.INDUCTION,
+];
 
 const convertExpression = (data: UniProtkbAPIModel) => {
   return convertSection(
