+++ src/model/uniprotkb/sections/FunctionConverter.ts
@@ -1,4 +1,4 @@
-import CommentType from '../../types/CommentType';
+import { CommentType } from '../../types/CommentTypes';
 import KeywordCategory from '../../types/KeywordCategory';
 import FeatureType from '../../types/FeatureType';
 import EntrySection from '../../types/EntrySection';
