+++ src/model/uniprotkb/sections/ProteinProcessingConverter.ts
@@ -1,7 +1,7 @@
 import FeatureType from '../../types/FeatureType';
 import KeywordCategory from '../../types/KeywordCategory';
 import EntrySection from '../../types/EntrySection';
-import CommentType from '../../types/CommentType';
+import { CommentType } from '../../types/CommentTypes';
 import { convertSection } from '../SectionConverter';
 import { UniProtkbAPIModel } from '../UniProtkbConverter';
