+++ src/view/uniprotkb/ExpressionSection.tsx
@@ -6,7 +6,7 @@ import FreeTextView from './components/FreeTextView';
 import KeywordView from './components/KeywordView';
 import XRefView from './components/XRefView';
 import { UIModel } from '../../model/uniprotkb/SectionConverter';
-import CommentType from '../../model/types/CommentType';
+import { CommentType, FreeText } from '../../model/types/CommentTypes';
 
 const ExpressionSection: FC<{
   data: UIModel;
@@ -19,11 +19,13 @@ const ExpressionSection: FC<{
     <div id={EntrySection.Expression}>
       <Card title={EntrySection.Expression}>
         <FreeTextView
-          comments={data.commentsData.get(CommentType.TISSUE_SPECIFICITY)}
+          comments={
+            data.commentsData.get(CommentType.TISSUE_SPECIFICITY) as FreeText[]
+          }
           includeTitle
         />
         <FreeTextView
-          comments={data.commentsData.get(CommentType.INDUCTION)}
+          comments={data.commentsData.get(CommentType.INDUCTION) as FreeText[]}
           includeTitle
         />
         <KeywordView keywords={data.keywordData} />
