+++ src/view/uniprotkb/PathologyAndBiotechSection.tsx
@@ -6,7 +6,7 @@ import KeywordView from './components/KeywordView';
 import EntrySection from '../../model/types/EntrySection';
 import hasContent from '../../model/utils/utils';
 import { UIModel } from '../../model/uniprotkb/SectionConverter';
-import CommentType from '../../model/types/CommentType';
+import { CommentType, DiseaseComment } from '../../model/types/CommentTypes';
 import XRefView from './components/XRefView';
 import VariationView from './components/VariationView';
 
@@ -22,7 +22,9 @@ const PathologyAndBiotechSection: FC<{
     <div id={EntrySection.PathologyAndBioTech}>
       <Card title={EntrySection.PathologyAndBioTech}>
         <DiseaseInvolvementView
-          comments={data.commentsData.get(CommentType.DISEASE)}
+          comments={
+            data.commentsData.get(CommentType.DISEASE) as DiseaseComment[]
+          }
           primaryAccession={primaryAccession}
         />
         <FeaturesView features={data.featuresData} sequence={sequence} />
