+++ src/view/uniprotkb/ProteinProcessingSection.tsx
@@ -6,14 +6,14 @@ import FeaturesView from './components/FeaturesView';
 import KeywordView from './components/KeywordView';
 import XRefView from './components/XRefView';
 import FreeTextView from './components/FreeTextView';
-import CommentType from '../../model/types/CommentType';
+import { CommentType, FreeText } from '../../model/types/CommentTypes';
 import { UIModel } from '../../model/uniprotkb/SectionConverter';
 
 const ProteinProcessingSection: FC<{
   data: UIModel;
   primaryAccession: string;
   sequence: string;
-}> = ({ data, sequence, primaryAccession }): JSX.Element | null=> {
+}> = ({ data, sequence, primaryAccession }): JSX.Element | null => {
   if (!hasContent(data)) {
     return null;
   }
@@ -22,7 +22,10 @@ const ProteinProcessingSection: FC<{
     <div id={EntrySection.ProteinProcessing}>
       <Card title={EntrySection.ProteinProcessing}>
         <FeaturesView features={featuresData} sequence={sequence} />
-        <FreeTextView comments={commentsData.get(CommentType.PTM)} includeTitle />
+        <FreeTextView
+          comments={commentsData.get(CommentType.PTM) as FreeText[]}
+          includeTitle
+        />
         <KeywordView keywords={keywordData} />
         <XRefView xrefs={xrefData} primaryAccession={primaryAccession} />
       </Card>
