+++ src/view/uniprotkb/components/ProteomesView.tsx
@@ -1,7 +1,7 @@
 import React, { Fragment } from 'react';
 import { InfoList } from 'franklin-sites';
 import { Link } from 'react-router-dom';
-import { Xref } from '../../../model/utils/XrefUtils';
+import { Xref } from '../../../model/types/CommentTypes';
 import { Property } from '../../../model/types/modelTypes';
 
 const ProteomesId: React.FC<{ id?: string }> = ({ id }) => (
